Add files via upload
This is my first web development project, Koekafe, a sleek and responsive website for a coffee shop. The website is designed with user-friendly navigation, featuring an elegant layout that showcases the cafe's menu, services, and contact information. Built using HTML5, CSS3, and JavaScript, it includes smooth animations, an interactive gallery, and a reservation system. The project demonstrates a strong foundation in web design principles and best practices for creating a visually appealing and functional website.